\name{gap.examples-package}
\alias{gap.examples-package}
\alias{gap.examples}
\docType{package}
\title{
Examples for genetic analysis package
}
\description{
The package contains several examples for use of the genetic analysis package (gap).
}
\details{
See various vignettes as listed by library(help=gap.examples).
}

\references{
Zhao JH, gap: genetic analysis package. Journal of Statistical Software 2007, 23(8):1-18
}
\keyword{package}
